Extracted helper function into common module
