+++ crates/sncast/tests/e2e/account/mod.rs
@@ -2,3 +2,5 @@ mod add;
 mod create;
 mod delete;
 mod deploy;
+mod helpers;
+mod list;
